+++ src/trogdord/config.cpp
@@ -14,6 +14,36 @@
 #endif
 
 
+// Maps env variables to config options they override
+const std::unordered_map<std::string, std::string> Config::ENV = {
+
+	{"TROGDORD_PORT",                             CONFIG_KEY_PORT},
+	{"TROGDORD_REUSE_ADDRESS",                    CONFIG_KEY_REUSE_ADDRESS},
+	{"TROGDORD_SEND_TCP_KEEPALIVE",               CONFIG_KEY_SEND_TCP_KEEPALIVE},
+	{"TROGDORD_LISTEN_IPS",                       CONFIG_KEY_LISTEN_IPS},
+	{"TROGDORD_LOGTO",                            CONFIG_KEY_LOGTO},
+	{"TROGDORD_INPUT_LISTENERS",                  CONFIG_KEY_INPUT_LISTENERS},
+	{"TROGDORD_OUTPUT_DRIVER",                    CONFIG_KEY_OUTPUT_DRIVER},
+	{"TROGDORD_REDIS_HOST",                       CONFIG_KEY_REDIS_HOST},
+	{"TROGDORD_REDIS_USERNAME",                   CONFIG_KEY_REDIS_USERNAME},
+	{"TROGDORD_REDIS_PASSWORD",                   CONFIG_KEY_REDIS_PASSWORD},
+	{"TROGDORD_REDIS_PORT",                       CONFIG_KEY_REDIS_PORT},
+	{"TROGDORD_REDIS_CONNECTION_TIMEOUT",         CONFIG_KEY_REDIS_CONNECTION_TIMEOUT},
+	{"TROGDORD_REDIS_CONNECTION_RETRY_INTERVAL",  CONFIG_KEY_REDIS_CONNECTION_RETRY_INTERVAL},
+	{"TROGDORD_REDIS_OUTPUT_CHANNEL",             CONFIG_KEY_REDIS_OUTPUT_CHANNEL},
+	{"TROGDORD_REDIS_INPUT_CHANNEL",              CONFIG_KEY_REDIS_INPUT_CHANNEL},
+	{"TROGDORD_DEFINITIONS_PATH",                 CONFIG_KEY_DEFINITIONS_PATH},
+	{"TROGDORD_STATE_ENABLED",                    CONFIG_KEY_STATE_ENABLED},
+	{"TROGDORD_STATE_AUTORESTORE_ENABLED",        CONFIG_KEY_STATE_AUTORESTORE_ENABLED},
+	{"TROGDORD_STATE_DUMP_SHUTDOWN_ENABLED",      CONFIG_KEY_STATE_DUMP_SHUTDOWN_ENABLED},
+	{"TROGDORD_STATE_CRASH_RECOVERY_ENABLED",     CONFIG_KEY_STATE_CRASH_RECOVERY_ENABLED},
+	{"TROGDORD_STATE_FORMAT",                     CONFIG_KEY_STATE_FORMAT},
+	{"TROGDORD_STATE_PATH",                       CONFIG_KEY_STATE_PATH},
+	{"TROGDORD_STATE_MAX_DUMPS_PER_GAME",         CONFIG_KEY_STATE_MAX_DUMPS_PER_GAME},
+	{"TROGDORD_EXTENSIONS_PATH",                  CONFIG_KEY_EXTENSIONS_PATH},
+	{"TROGDORD_EXTENSIONS_LOAD",                  CONFIG_KEY_EXTENSIONS_LOAD}
+};
+
 // Default ini values
 const std::unordered_map<std::string, std::string> Config::DEFAULTS = {
 
@@ -220,6 +250,12 @@ void Config::load(std::string newIniPath) noexcept {
 	iniPath = trogdor::trim(newIniPath);
 	ini.clear();
 
+	// Start by populate the ini object with its default values
+	for (auto &defaultVal: DEFAULTS) {
+		ini[defaultVal.first] = defaultVal.second;
+	}
+
+	// Next, override the defaults with any options in the ini file
 	if (
 		0 != iniPath.compare("") &&
 		STD_FILESYSTEM::exists(iniPath) &&
@@ -257,11 +293,11 @@ void Config::load(std::string newIniPath) noexcept {
 		}
 	}
 
-	// Populate the ini object with defaults for any values not set in the ini
-	// file.
-	for (auto &defaultVal: DEFAULTS) {
-		if (ini.end() == ini.find(defaultVal.first)) {
-			ini[defaultVal.first] = defaultVal.second;
+	// Finally, override any settings that have been configured in environment
+	// variables (these take precedence over trogdord.ini.)
+	for (auto &envVar: ENV) {
+		if (const char *setting = std::getenv(envVar.first.c_str()); setting) {
+			ini[envVar.second] = setting;
 		}
 	}